Fix and enhance addprinc type
Add path so that puppet does not complain: "... is not qualified and no
path was specified."

Support use of kadmin as well as kadmin.local distinguished by the local
property. Support use of ticket cache as well as keytab for kadmin.
Support retry in case kadmind is restarting or similar.

Do not create principals with insecure default password. Instead use
-nokey to create them without a key.

Do not require the KDC to be running but the database to be created for
kadmin.local. Do include the kadmind class to make sure the package
containing kadmin.local is getting installed.