Add types for keytabs and ticket caches
Add types that support the creation of keytabs, adding keys to them
using ktadd and acquiring tickets and putting them in tickets caches.
Keys can be exported using either kadmin.local or kadmin. A large number
of properties can be specified as to who the files should belong to and
whether they should be recreated each run or left alone once created.