Fix up creation of and dependencies on database directory
Refer to the database directory by path instead of title because it was
done inconsistently before and the path might be the same as the
kdc_conf_dir and the collision can not be detected and avoided
otherwise.

Do not force a mode of 0700 because it's not necessary (the KDC secures
file individual files' modes itself) and might cause flapping depending
on which file resource (kdc_conf_dir or kdc_database_dir) "wins", i.e.
is instantiated.

Make kpropd depend on the database directory as well because it does
actually need it to exist in order to be able to put the just replicated
database into it.