Implement regular key changes
Extend the ktadd type to use a helper script to check for key expiry and
renewal. Also takes care of leaving old sets of keys in the keytab as
long as there might be tickets around still encrypted with them. Can
optionally create keytab/principal-specific helper scripts to be u sed
for individual checks such as usually implemented via monitoring.